Seperate header from Main
Seperate Employee header from Main
